+++ CHANGELOG.md
@@ -13,6 +13,8 @@ This is a template for a release section.
 
 ### Added
 
+- Solving quadratic inequalities by finding the roots and considering the inequality sign (PLUT-1020)
+
 ### Fixed
 
 - Exclude HPA from development. (INF-21)